Sync & Correlation.
During sync and correlation = db,
an attachment or malware did not get processed into
Attribute.data, so will not be synced.
Now, conform other correlation methods being 'default' or 'sql'
the attachment or malware is synced as well.
(master has been synced with mil.be not using db correlation,
so should have the data.)