Better placement of plugins (touching RBAC & Audit log)
If it's just an existing behavior or lib,
place it in a plugin directory structure in <cydefsig>/plugins.

If there is a need to change an extern existing plugin,
extend the existing plugin by a new plugin in <cydefsig>/app/Plugin.

This way there is a very clean devision between own and external code.
The external code can be updated without touching own nor changed code.