event/attribute delete
In version 1 and 2 of misp/cydefsig there's a delete button upper left
in the menu that a) does not delete or b) does not return to a visable
url after deletion.
As a 'fix' those delete buttons are now removed, given there does still
exist delete in the index view.