Version 1.0 of MISP XML Document Type Definition.
The first version of the XML format is loosely based on the current XML
format used by MISP in commit 84b552fb7441bf2beb0c711acde3b0af336afba8.

The purpose is to track down the changes in the format and especially
to ensure a consistent definition of the XML format for external tools
and software using the MISP XML format.