Moved attachments into a separate folder
This pull request will create a subdirectory for attachments, which provides separation of attachments from the other directories.

A feature I wish to implement is the option to use s3fuse on the attachments folder. However, currently the attachments are created as eventId amongst the other directories within files preventing me from use fuse easily.