fix: skip the import of mixbox for users of older stix libraries
If you rely on old idgen from previous stix libraries, mixbox is not installed.
This completes the fix #2186 and should be fine for old and new stix libraries.

A partial lyric has been included in this commit to ease the pain to work ##comma##:

Money for nothin' and your stix for free
Money for nothin' and stix for free