new: [updates] Added lock mechanism
This commit introduce a lock mechanism preventing the start of multiple
update instances.
Basically, when an update starts, it records the start time of the
update; when the update process ends, it clears the entry.
Whenever a new update process is started during an ongoing update, it
will check the start time against the configuration threshold
(MISP.updateTimeThreshold) and continue if allowed.