chg: remove centos step
The instruction to enable the SCL repos is already provided in step 1.4,
and the package centos-release-scl does not exist on RHEL7. The fact
that step 1.4 may not work on a trial subscription is not relevant
because assumptions 0.1 and 0.2 specifically state that a support
agreement is required to be able to complete the installation.