chg: [events:index] Quick search in event index only searches on event.info
The quickfilter used to search in all possible fields in the event and it was massively used by users who only wanted to search in the event.info field.
The search_all feature has been moved in the search index modal under the name `search in all fields` and the quick search input now only perform searches on the event.info field.