Fixes issue #6379
The NIDS export creates two rules for attributes with type 'email' (a
src and dst rule). However, the same SID was used for both rules. Since
SIDs must be unique for a ruleset, this will be logged as an error by
Suricata and the rule is not loaded (see issue #6379).

This fixes the issue by incrementing the SID before creating the second
email rule.