new: [attribute type] `dom-hash` is a structural fingerprint of HTML's Document Object Model
`dom-hash` is a structural fingerprint of the HTML's Document Object
Model (DOM) originaly developed by CERT.PL

The fingerprint is calculated by extracting all the tag names (ignoring
the content and attribute of the HTML Page). The tag names are
concatenated with a pipe value `|`, hash with the SHA-256 algorithm
and truncated with the first 32 characters.

Software like LookyLoo[1] implemented the algorithm which can be used
in MISP to share and correlate information about similar web pages (e.g.
phishing pages).

[1] https://github.com/Lookyloo/lookyloo/commit/466a3c56148f2ddb911620fd24e4f0c9d602a6a3